backend/requirements.txt
fastapi
uvicorn
python-socketio
scikit-learn
numpy
